Make DimensionNameError propagatable
DimensionNameError is now a subclass of ButlerUserError so that it can be sent from server to client.  It now also subclasses KeyError to replace a  case in DataCoordinate.standardize where we were throwing KeyError with the same meaning as this exception.
